+++ product-server/src/main/java/com/product/sysuser/service/ISysUserService.java
@@ -6,9 +6,16 @@ import com.product.sysuser.bean.SysUser;
 
 public interface ISysUserService {
 
+	/**
+	 * 保存用户信息:
+	 * <br>1.主键为null则执行insert
+	 * <br>2.主键不为null则执行update
+	 * @param sysUser
+	 * @return
+	 */
 	int save(SysUser sysUser);
 
-	SysUser findSysUser(Integer id);
+	SysUser findSysUser(Long id);
 
 	List<SysUser> findSysUsers(SysUser sysUser);
 }
